Move pkg/kp underneath pkg/store and rename to consul.
pkg/kp is inextricably tied to Consul semantics. A lot of the operations
are specific to consul, e.g. session creation, locking, and accepting a
PodPrefix parameter in GetPod() and SetPod().

This PR organizationally paves the way to support multiple backends by
naming kp what it is: consul-specific.

The idea is that pkg/store will have constructors for separate backend
objects that provide the same functionality. For example, any package
that needs to operate on replication controllers might define an
interface like

type RCGetter interface {
    GetRC(rc_fields.ID) rc_fields.ReplicationController
}

which it will take as an argument to its constructor.

pkg/store can offer multiple constructurs e.g. for ConsulStore and
ETCDStore, both of which implement the GetRC interface.