Include directory names in entry point name for runit services.
Previously an entry point like bin/launch in a launchable called
"banana" within a pod named "apple" would generate a runit service named
"apple__banana__launch" for hoist launchables, the "launch" portion
coming from taking the basename of the full path to the entry point.

This always posed the problem of specifying multiple entry points like
bin/start and script/start resulting in a service naming collision, an
error that the preparer appropriately detected.

This commit changes this behavior to include the full relative path to
the entry point in the runit service name as well as the $ENTRY_POINT
environment variable. This allows multiple entry points to share a
filename and additionally is easier to compute for external systems that
may wish to observe the exit code of an entry point within a pod.