Add ability to repair the workspace file in pod finish reporter.
In order to publish information about the exit status of processes
started by a uuid pod, a runit hook is written that writes information
about each process exit to a local sqlite database. The preparer has a
"pod process reporter" which routinely polls this database for new
entries and writes them to consul.

To track the highest ID it has already seen in the database, it writes
the ID to a file at the end of each processing loop.

Sometimes we see the file end up with 0 bytes in it which means the pod
process reporter gets an error on every loop and can't do its job.

This commit adds the ability for the reporter to repair the file by
looking up the last ID in the database and writing that to the file.
This means that some finish data may be lost, but the alternative is to
double submit exit information to consul that might be old, so this
scheme is preferable.