Add :sync_consul_deps rake task for upgrading consul dependencies.
Consul is probably the vendored largest dependency in the P2 repo, and
commonly needs to be upgraded. This is a challenge because consul does
not use godep (it uses govendor).

Furthermore we might share dependencies with consul, and we want to
upgrade our dependencies to match theirs to have the highest confidence
in our build while not downgrading any of our packages.

The rake task iterates over the dependencies vendored in the currently
checked out consul version, makes sure they're in $GOPATH, checks if
consul has a newer vendored commit, and if so will check out that commit
and perform a godep update.