Restore old naming scheme for runit service directories for legacy pods.
Commit b22751cc08446367b6181ee5b602fc0d3a469c0e had modified the naming
scheme for runit service directories (i.e. /var/service/*).

Prior to that commit, the directory names were built like the following:
"/var/service/{podID}__{launchableID}__{filepath.Base(entryPointPath)}
yielding "/var/service/some-pod__some-launchable__launch" for a pod
named "some-pod", launchable named "some-launchable" with the bin/launch
entry point used.

After that commit, the naming scheme was modified to include information
about the directory hierarchy to the entry point relative to the root of
the artifact: e.g. bin__launch is used for the entry point portion
instead of just the file basename. This allows artifacts to have
multiple entry points with the same filename (in different directories).

This commit reverts this naming scheme change for "legacy" pods which do
not have a UUID, and keeps the new naming scheme if the pod has a UUID.