ds: failsafe: If no nodes are eligible, do nothing
It is expected that all daemon sets have non-empty target node sets at
all times.

If our label store ever indicates otherwise, we assume the label store
has erroneously returned empty results, and do nothing, to be safe.

This is similar to the p2-preparer's failsafe where it refuses to act if
it does not see its own key in intent.

The proper way to unschedule a daemon set on all its current nodes is to
delete it.