Run bin/enable even when there are missing entry points.
Historically, the hoist artifact spec was rigid in its requirement that
any services that should be started under runit must be at or under the
bin/launch location in the artifact. Under this scheme, it was not
considered an error for an artifact to not have bin/launch, it would
simply be downloaded, and have bin/enable run after starting no
processes.

P2, however, has added the option of specifying alternate entry points
within an artifact in your pod manifest. At this point it became an
error if entry points were missing (including the implicit bin/launch if
the option is not overridden). This had the side effect of causing
bin/enable to not be run for apps that do not start any services.

This commit returns to the behavior of running bin/enable if the only
startup error was related to missing entry points. Other errors will
still cause bin/enable to not run.