Add retry behavior to p2-finish-env-extractor.
p2-finish-env-extractor is a binary that p2-preparer wires into runit to
capture the exit code of the processes that runit manages. In the case
of uuid pods p2-finish-env-extractor writes a row to a sqlite database
on disk to capture the exit code.

This scheme can result in lost information because writing to a sqlite
database always locks the entire database, so if multiple processes exit
at the same time they will compete for the ability to write to the
database.

This commit adds the ability to make p2-finish-env-extractor retry
errors writing to the database to improve the likelihood of a row making
it into the database.