Make unmarshaling an RC with missing replicas_desired an error.
This adds to the safety of the RC farm, because it allows us to
distinguish a JSON RC missing the replicas_desired field completely from
one that has it but a 0 count for the value.

It is an error to unmarshal an RC that does not have a replicas_desired
field.

This protects us against a scenario where the json tag might be changed
such that all stored RCs will have replica counts of zero. There is
already a failsafe in place that prevents the farm from taking action of
the sum of all of the rc replica counts is zero, but this adds an
additional layer of safety.