Make RC farm watch keys in consul (not values) to reduce bandwidth.
Bandwidth to and from the leader of the consul cluster is P2's
scalability bottleneck. The RC farm is an easy target for optimization
since there can be 1000s of RCs and a watch is performed on the entire
RC subtree, which changes a lot as every rolling update triggers
serveral changes to replicas_desired counts within RCs. This is
exacerbated when there are multiple servers running RC farms for fault
tolerance and redundancy as they will all be executing watches.

The RC farm is structured such that there is a top level goroutine
watching the entire RC subtree in consul whose job is to spin up
goroutines to handle RCs. Each of these sub-goroutines executes its own
individual watch on the RC it handles which is good from a bandwidth
standpoint as queries aren't returning as often and the data that is
returned is small.

The top-level goroutine doesn't really need the values of all of the RCs
returned, it just needs the keys to determine which have been deleted
and which RCs are new. Watching keys instead of values means that the
data returned by each watch will be smaller which should reduce outbound
bandwidth from consul.

The one behavioral tradeoff that had to be made is in one of the
failsafe mechanisms. Prior to this change, every time the RC watch
returned a sum was taken of the replicas_desired counts of every RC. If
that sum was zero (pathological case) the server would panic to prevent
the farm from uninstalling a lot of pods when it shouldn't. Now this
check is performed at startup by a one-time fetch of all of the RC
values, and subsequently the only failsafe protection is against all of
the RCs disappearing.

This tradeoff is deemed acceptable because another layer of safety was
added in the JSON unmarshaling of RCs; it's now an error to unmarshal an
RC from a json structure that does not have a replicas_desired field,
which is what the original failsafe was designed to prevent (for
instance if the json key was renamed and the RC json blobs in consul
weren't similarly migrated).