Don't run disable/enable for oneoff pods.
oneoff pods don't need to have their disable/enable scripts run because
they are oneoff jobs with immutable state. Furthermore running
disable/enable scripts may impact other pods with the same pod ID on the
same physical hardware depending on the implementation if
enable/disable.

This commit creates a (artificial at this point) distinction between
uuid pods and oneoff pods. A uuid pod is a pod whose data is stored in
the pods/ consul tree with an associated uuid. A oneoff pod is a pod
that is meant to run once and does not represent a repeatedly running
service like a web server.

These two concepts are tied together currently but now there is a
IsOneoff() function on the hoist launchable type which may change at
some point.

This commi