Rename "p2-pcctl update" command to "p2-pcctl update-annotations".
This is in anticipation of adding a command for updating the pod
selector of a pod cluster. It is unlikely that an operator will want to
change annotations and a pod selector at the same time so having two
different commands makes things easier.

Furthermore, the help text on update annotations operations was
clarified, some code comments were added, and multiple var() statements
better group the flags for each command.