Rename selectorFrom() to defaultSelector() in p2-pcctl.
It's possible and occasionally desirable to be able to specify an
arbitrary pod selector for a pod cluster. In anticipation of adding a
command that does just that, the selectorFrom() command has been renamed
to indicate that it returns a suggestion/convention for a pod cluster's
pod selector but it is possible to change later.