Add update-selector command to p2-pcctl.
This command allows you to update the pod selector on a pod cluster,
effectively changing its membership. Since this is a potentially
dangerous operation that can have effects on things like service
discovery, a diff is printed showing the pods that would gain and lose
membership to the cluster, asking the operator to confirm before
applying the update.