Remove Store interface from pkg/store/consul/dsstore package.
Rationale:

1) We want to move to datastore-agnostic APIs to allow us to switch
backends. Today P2 code is very coupled to consul, and we have many
store interfaces that use a consul client and expose interfaces that are
imported into numerous packages. Gradually we want to split these
interfaces apart into datastore-agnostic ones.
2) Fat interfaces are an anti-pattern. If the store packages simply
return structs that are passed around our code as interfaces, it's easy
to tell which functionality of the store is used in which parts of the
codebase. It also means that the interface implementations are easier to
swap to different implementations or mocks for tests.