Split up the DaemonSetStore interface in pkg/ds.
Now there is a new interface called DaemonSetLocker which contains the
functionality necessary for a farm instance to claim a distributed
exclusive lock on a daemon set before processing it.

This change is useful because DaemonSetLocker is now devoid of consul
semantics (although the only real implementation of it requires a consul
client). This paves the way for implementing it with other backend
types.