Remove Store interface from pkg/store/consul/rcstore package.
This is a step toward building alterative storage implementations to
direct consul access. The rcstore.Store interface was very wide and
defined some operations that are very tied to consul (e.g. use a consul
session) and others that perform CRUD style operations on types defined
by P2 which makes them easy targets for other implementations.

This change replaces the big interface with minimal ones throughout the
code base where RC stores are used.

There are some cases where a single instance of an rc store was required
and now there are up to 3 because the store object is distributed to
different packages and structs that have distinct interfaces. This is
considered desirable because of the win of being able to swap out small
interface implementations even though it results in passing the same
underlying value multiple times to the same constructor.