Make it easier to delete a uuid pod for which pod ID is not known.
A pod's uuid contains all the information necessary to delete that pod.
For safety, p2-rm historically required you pass the uuid as well as the
pod ID to confirm that the user is deleting the pod they expect they
are.

However this makes it harder to use the tool. This commit adds a
confirmation prompt that tells the user the pod ID of the pod being
deleted to preserve safety while adding some ease.