Return special and testable error type when no labels are found.
Consul has been known to return spurious 404s for data that is known to
exist. As a result it makes sense to design our systems to discard
results for queries where data is expected to exist under all
conditions.

This commit makes the label applicator return a special NoLabelsFound
error when listing all labels for a particular type. There are certain
label types (e.g. replication controllers) for which we expect there to
always be labels. In this case clients seeing this error should discard
the results rather than attempt to use an empty label set.

This also affects the GetMatches() call on consul applicators because it
calls ListLabels under the hood and returns non-nil errors.