Fix callsites of GetMatches() to use new label type where applicable.
Most callsites of GetMatches pass labels.POD and labels.NODE as the call
sites. We expect there to always be labels present for those types
during normal operation of P2, so it's good that these call sites will
now see errors when calling GetMatches if no labels of those types were
found in consul.

p2-label however takes a freeform label type on the command line, so now
it prints different output to the user if no labels were found for the
particular type given.

Furthermore, the label HTTP server will now return a 404 error if the
select endpoint is called for a particular label type and selector and
no labels were found for the given label type. This is good since it
allows clients to distinguish a selector matching 0 results from there
being 0 results to begin with.