Fix callsites of ListLabels() to use new error type.
The only caller of ListLabels() is within the label http server. Now it
will return a 404 error instead of an unavailable error if the
underlying error was missing labels for the given label type. This will
let callers distinguish between transient server failures and failures
due to missing or apparently missing data.