Make pod argument to p2-rm optional for uuid pods.
Commit 67002b46ed79 intended to make p2-rm easier to run on oneoff pods
by making the pod ID argument optional and showing a confirmation prompt
instead. However that commit failed to remove the kingpin requirement on
the pod argument, making p2-rm fail if the argument was not provided.

This commit makes it required only for deleting legacy pods as it should
be.