Deprecate wait index on WatchPodStatusRequest for gRPC.
The grpc pod store provides a gRPC endpoint called WatchPodStatus which
streams updates to a pod status record back to the client using consul
wathes on the backend. The responses return the LastIndex of the consul
query that returned the result, and the client may pass an index to use
in the request. This mechanism allowed the client to guarantee it won't
get a result that is more stale than the last one it got.

However this mechanism has the downside of exposing consul indexes to
the client, which couples our API to a particular data store.

This commit provides the desired guarantee that each successive result
is newer than the last even across separate rpc calls by first doing a
consistent fetch from consul to seed the index to use for subsequent
watches.

It no longer reads the WaitIndex from the request in preparation for
that parameter being removed.