Restructure podstore grpc client constructor to take a *grpc.ClientConn.
Previously it took parameters required to Dial() to create a connection.
This complicates using the client if you want to use complex options
such as various security options or load balancing.

Furthermore this change makes the user of the client responsible for
closing the connection which simplifies the client code.