Split podStatusResultToResp function into usable parts.
Previously it was taking a struct with a PodStatus field and an error
field and returning a pod status response and an error. These two
arguments and return types do not affect each other so it made sense to
split these into separate functions to make them easier to use.