Modify helper function to convert consul paths to pod ID if applicable.
Currently p2 allows the intent/ and reality/ tree to contain "uuid" pods
and "legacy" pods. Using intent as an example, a uuid key might look
like intent/example.com/ffaaca6b-106d-466c-8ffc-a626dfa4e6a4 while a
legacy key might look like intent/example.com/mysql. There was a helper
function called PodUniqueKeyFromConsulPath() that takes as an argument a
consul path and returns a pod unique key (uuid) if the key is formatted
appropriately. PodID was not needed from this function because typically
that could be deduced from the value of the key by calling manifest.ID()
on the manifest contained in the key's value.

In anticipation of adding some bandwidth-reducing measures that touch
the keys but not the values of intent pods, it is now useful to be able
to deduce the pod ID from the key for legacy pods.