Add WatchIntentKeys function to consul store.
This function can be used by bandwidth-conscious clients who want to
receive quick updates to intent keys when they change but wish to avoid
unnecessary fetches of values. Keys are returned on a channel along with
information about consul's ModifyIndex for the key which allows clients
to cache values and only request values when the modify index changes.