Make the default consul wait time configurable in p2-preparer.
In large deployments of P2, a lot of bandwidth from the consul server is
used by potentially thousands of preparer instances pulling manifests
that might be multiple Kb constantly. When no intent data for a node is
changing, this means a request completes every 5 minutes, returning the
full data.

This commit makes it easy to change the wait time on consul watches
performed by p2-preparer so that bandwidth usage during periods of
steady data can be tuned.