Unconditionally install hooks on preparer startup.
Prior to this commit the hook manifest (fetched from preparer manifest
config) was compared to the current manifest written to disk for hooks.
If the manifests were the same, hook installation was skipped.

This commit removes this check and just unconditionally installs the
hooks.

This is for two reasons:
1) Installing is low-cost anyway. We don't redownload artifacts if
they're present on the disk, and the rest is just linking and writing of
the current manifest file.
2) It's possible that the hooks might have been removed but the
current_manifest.yaml file for them was untouched, which will cause the
preparer to incorrectly not install them.