Set the "last" symlink even if there is an error stopping a launchable.
The "last" symlink points to the launchable version that was previously
running. It is used by hoist launchable pruning code to preserve the
last launchable to make sure that rolling back a pod doesn't require
re-downloading the launchable from the network.

A launchable which has a stop error such as missing entry points (which
is sometimes the desired state for a pod) was not having its last
symlink updated, and therefore a very old launchable install was being
preserved by pruning code.