Ask for confirmation before changing RC replica count in p2-rctl.
The p2-rctl set-replicas command is very useful and powerfull, however
it's easy to make a typo and cause a problem. If the replica count being
set is lower than the currently set one, it will cause pod
uninstallations.

This commit adds a confirmation prompt which will give a user time to
confirm that the count they typed in was in fact what they wanted.