Add TransferReplicaCounts function to consul RC store.
TransferReplicaCounts() "transfers" a number of replicas from one RC to
another utilizing consul's transaction functionality. This allows
invariants to be maintained about the total of the replica counts of two
RCs even across consul server failures.

An integration test was added to test this functionality since it is
hard to implement a fake consul server that has the full transaction
capability.