Make rolling update loop use rcstore's TransferReplicaCounts()
This makes the updating of RC counts during a rolling update
transactional which means if the p2-master server or consul cluster
faults the invariant of RC count totals will not be violated.

This change notably removes the old check-and-set functionality on
updating RCs, but given that the RC mutation locks are held during a
rolling update we have reasonable confidence that the RC counts were not
mutated underneath the rolling update loop.