Allow tuning of request rate by WatchService
WatchService performs a repeated list on v1/kv/health/{serviceID} using
a wait index. This means that in periods of little change for the
service's health, we will be getting a request about every minute
(because watches expire after 5 minutes even if the data doesn't
change).

However, if the data is changing frequently, the upper bound on QPS is
1/second. In the case of daemon sets for which there might be thousands
of health keys, it's quite easy to end up in a scenario where health is
changing very frequently and QPS is close to that upper bound.

This commit adds a watchDelay parameter to WatchService which guarantees
that no two List() operations will be initiated within that interval.