ds: spawn child when acquiring lock on updated DS
If we have just acquired a lock on a daemon set, necessarily it is not
one of our children. Therefore, any attempt to send on its updated
channel will trigger a nil dereference. This code has never worked.

Attached test demonstrates, though "Attempt to acquire unchanged daemon
sets" of #813 alleviates the problem slightly. With #813, the attached
test only sometimes dereferences nil because sometimes the lock is still
held when the second farm attempts to access and the daemon set moves to
the "unchanged" or "same" category.

A similar mistake was fixed in #753.