ds: Attempt to acquire unchanged daemon sets (#813)
Without this change, if a farm crashes, no other farm will pick up the
daemon sets managed by the former farm.

This means that if a daemon set gains a new node, nothing will schedule
onto the new node.

Attached test fails without attached code change.

Note that this is a very low-effort way to effect this change. The code
could be restructured to look more like an RC watch and get rid of
WatchDiff, but this requires more time to work with the code.