Remove replication controller watchers from roll farm.
As a result of splitting up the giant preemptive *.Store interfaces, it
became clear that the rolling update store was passing a store that
watches RCs just to instantiate a ReplicationController struct which
wouldn't even use that parameter.

Now that it is completely unused, we can delete it from the roll farm
completely.