replication: Stop health watch when done (#816)
The AggregateHealth was never being stopped. Luckily, since nobody reads
off the channel, it blocks sending on the channel. So we leak
goroutines, but at least it's not as bad as continually hammering
Consul.

This is safe: The only use of aggregateHealth is as an argument to
updateOne, and updatePool will stop the function from returning (and
Stopping the aggregateHealth) until all updateOne have finished.

First discovered in:
https://github.com/square/p2/pull/738#discussion_r98260867