Add a test that health writes to consul are retried after failures.
This commit contains a test that was an attempt to reproduce a bug seen
in production where occasionally a preparer will fail to sync its local
state in consul. This bug tends to be accompanied with "write denied"
error messages from consul, suggesting that the health manager loop was
not properly retrying a write after a failure. The test did not actually
uncover a bug but it's probably useful to commit anyway.