raise log level of health manager from Debug to Info
This shouldn't make the logs too much more noisy: there will be
additional logging when the session changes (which shouldn't happen
except when restarting under normal circumstances) and we'll see a few
extra lines for each health write.

This should help us debug reported incidents of health getting "stuck"
in the preparer, where it fails to sync a state change to consul.