turn off fail fast call option for gRPC retries
A common pattern in P2 code is to call a function that returns a channel
of results, the values on which are the result of calls to Consul.
Errors connecting to consul are typically swallowed and retried.

When providing a gRPC implementation of this same interface, a similar
approach was taken where an error making the gRPC call results in
continuous retrying forever. In the event of a broken TCP connection,
these RPC calls would fail immediately, resulting in a lot of log
messages.

This commit disables the "fail fast" call option when retrying an rpc,
which means that the call will wait until a TCP connection is ready
before returning. This will reduce the amount of logging when the TCP
connection breaks, but doesn't simplify the code because there are still
certain errors that could occur, so a retry loop is still necessary.