fix a race in pkg/replication
Each replication struct has a enactedCh field which is a channel that
will be closed when the replication is done. It is initialized to a
closed channel, and then overwritten as an open channel when the
replication starts. This commit adds a mutex that synchronizes access to
the channel to prevent data races.