fix more races in pkg/ds
Many places in the daemon set farm and the farm test access a map of
daemon set ID to a daemon set struct. There was a mutex that seemingly
was intended to be used to synchronize access to the map, but was not.
This commit makes sure all map access is done while holding that mutex.