fix map-access race in fake daemon set store.
The daemonsets map maintained by the daemon set store now has its access
synchronized properly by a mutex.

prior to this commit, WatchList() in the fake ds store contained a tight
for loop calling List(), which results in a lot of log spam and a busy
CPU. Now mutation operations will write to a channel if present to tell
the watch that it should List() again.