fix a data race in pkg/logging
A global counter is maintained for logging purposes. The atomic package
was used to make incrementing of the counter atomic. However, the
counter was stored on the processCounter struct itself, and access to
that struct was not synchronized. This problem was fixed by making the
counter a package global variable and making the struct simply refer to
it instead of a field on itself.