Fix data race in health manager test.
The health manager was using a global variable to determine the retry
time between failed health writes. A test that was testing these retries
was overwriting this global variable to reduce retry time in tests,
causing a data race.

Now that retry time is explicitly configured on the struct that does the
retries.