add Create() to audit log store
Unlike other Create() methods on other store types, this one does not
need access to a consul store to work. You pass it a *api.KVTxnOps
object and it will append an operation that will create the appropriate
audit log message. It is the responsibilty of the caller to apply the
transaction using a consul client.