make the roll store use transactions on creation operations
First of all, this simplifies a lot of code within the roll store and
also makes it safer. The roll store attempts to grab locks on both of
the new and old RCs, and in some cases needs to also create one or both
RCs. Using transactions makes it possible to create all required
entities and also check that the locks are still held while creating the
new RU, and have all or none of those operations succeed together.

A transaction must be passed into the create operations in anticipation
of callers wishing to batch up RU creation with other transactional
creations, such as audit log records.