use commit hooks in roll store to avoid committing transaction
Prior to this commit, the roll store would commit the transaction that
was passed to any of its many Create* functions. This was a necessity
because part of the transaction is checking that some locks on the RCs
are held, which will only be true if the session that created the locks
is not yet destroyed, however destruction was happening in a defer
session.Destroy().

This commit instead registers the destruction of the session as a commit
hook on the transaction, so that the transaction can be committed by the
caller into the roll store instead of the roll store itself. This is
useful because the caller may wish to add more operations to the
transaction after the creation of the RU.