Use a map instead of linked list to track watchers in consul aggregator.
The consul aggregator within pkg/labels is designed to allow clients to
register for updates to the membership if a label selector without
increasing load on the consul database. The aggregator tracks many
watchers and their selectors, and for each label update from consul it
computes the matches for every selector and passes the results on the
watcher's output channel.

A linked list was being used prior to this commit due to the fact that
it wasn't expected for there to be many watchers at a given time.
However now that we're going to allow subscribing to watch a selector
over gRPC, the gRPC server may in fact accumulate thousands of watchers,
and iterating a linkedlist in order to de register watchers while
holding a mutex is inefficient.

Furthermore, we expect that there may be multiple clients that watch the
same selector, and we wish to deduplicate the work required to service
this in the aggregator, which will require quick ability to look up a
watcher based on the label selector.