consul aggregator: make duplicate selector watches efficient.
If two separate clients of the consul aggregator set up watches on the
same selector, we don't want to compute the selector matches twice. This
is a realistic scenario for the gRPC label store server to encounter if
there are a lot of clients all watching the same selector.

To address this, this commit changes the data structure in which
watchers are stored in the consul aggregator. Now the map from label
selector string -> watcher has been modified to be a map from label
selector string -> map[watcher]struct{}. This way the result can be
computed once and then passed on the result channel of every watcher.