Export function in daemon set farm to identify contention.
During normal operation of the daemon set farm, it watches for new and
updated daemon sets and then checks if the daemon set contends with any
preexisting daemon sets where contention means both selectors overlap by
at least one node.

This function was private and therefore couldn't be used in other code,
e.g. in p2-dsctl in which it might be useful to prevent contending
daemon sets from entering consul in the first place.

This commit also addresses a bug in which only daemon sets that are
children of the current farm instance were considered for contention,
whereas the entire set of daemon sets should be used.